Ensure correct order of character replacing
The for..in operator is not guaranteed to work in the same order across
different browsers. The correct order of character processing however is
important for the parsing process since the HTML tags that will be inserted
into the string can contain PIML markdown characters. Therefore a character
like '/' which occurs in all HTML tags must be processed before any other
character.